Axis: Work around "homing done when hitting a limit switch" in MR
Whenever homing, don't update the limit switches, because that
would make MR to think that the motion is completed.
DMOV will go high, and the motor continues homing.
Note: When starting to home when homig from a limit switch, we hide
that as well, as soon as home() is called.

https://github.com/epics-modules/motor/pull/19